app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
html2text==2024.2.26
PyJWT==2.10.1
sqlalchemy==2.0.36
alembic==1.14.1
aiosqlite==0.20.0
psycopg2-binary==2.9.10
cryptography==44.0.0
bcrypt==4.2.1
email-validator==2.3.0
